changed order of beacon recognition in parser
I encountered a problem, where after a short time, only iBeacons got detected. I digged into the code and saw, that while the rule for detecting iBeacons was valid, the rule for detecting eddystone beacons was also valid. This means, that it could have been both, ibeacon and eddystone. But as far as the eddystone identification uses a UUID, which should be unique, i think, beacons having this uuid should not be misidentified as ibeacons. I simply changed the order, that whenever the uuid is valid, it is a eddystone beacon. Otherwise it can also be an iBeacon.